+++ ClimMob.R
@@ -5,13 +5,14 @@
 # # ................................................................
 # # ................................................................
 
-tag <- "nic_apante_2015"
-args <- c(paste0("dev/data/",tag,"/data.json"), paste0("dev/data/",tag,"/info.json"),
-         paste0("dev/output/",tag,"/"), "TRUE","en","docx",
-         "participant","item", getwd())
+# tag <- "nica_apante_2015"
+# args <- c(paste0("dev/data/",tag,"/data.json"), paste0("dev/data/",tag,"/info.json"),
+#          paste0("dev/output/",tag,"/"), "TRUE","en","html",
+#          "participant","item", getwd())
+# source("dev/run_climmobv2.R")
 
 # get the arguments from server's call
-# args <- commandArgs(trailingOnly = TRUE)
+args <- commandArgs(trailingOnly = TRUE)
 infoname    <- args[1] # a json file with parameters for the analysis
 outputname  <- args[2] # a json file with the results
 pathname    <- args[3] # the path where results will be written
@@ -39,6 +40,8 @@ library("pls")
 library("gtools")
 library("ggplot2")
 library("igraph")
+library("mapview")
+library("ggrepel")
 
 source(paste0(fullpath, "/R/functions.R"))
 
@@ -53,8 +56,6 @@ pars <- ClimMobTools:::.decode_pars(pars)
 cmdata <- jsonlite::fromJSON(outputname)
 class(cmdata) <- union("CM_list", class(cmdata))
 cmdata <- as.data.frame(cmdata, tidynames = FALSE, pivot.wider = TRUE)
-
-source("dev/run_climmobv2.R")
 
 # ................................................................
 # ................................................................
@@ -115,7 +116,7 @@ missexp <- 0.8
 minsplit <- ceiling(nrow(cmdata) * 0.1)
 if (minsplit < 10) {minsplit <- 10}
 # Set alpha
-sig_level <- 0.05
+sig_level <- 0.1
 
 # method for adjustments for confidence intervals and setting widths for comparison. 
 # Defaults to B-H (Benjamini an Hochberg). Any of the methods from p.adjust will work here 
@@ -141,6 +142,7 @@ info_table_typeinfo <- "" #info.table.typeinfo <- "expert advice"
 # ................................................................
 # ................................................................
 # Run analysis ####
+dir.create(pathname, showWarnings = FALSE, recursive = TRUE)
 source(paste0(fullpath, "/R/analysis_climmob.R"))
 
 # ................................................................
@@ -155,7 +157,6 @@ output_format <- ifelse(extension == "docx","word_document",
                         paste0(extension,"_document"))
 
 # produce main report if output type is "summary" or "both"
-dir.create(pathname, showWarnings = FALSE, recursive = TRUE)
 rmarkdown::render(paste0(fullpath, "/report/", language, "/mainreport/mainreport.Rmd"),
                   output_dir = pathname,
                   output_format = output_format,
@@ -165,3 +166,4 @@ rmarkdown::render(paste0(fullpath, "/report/", language, "/mainreport/mainreport
 # if (infosheets) {
 #   source("Farmer Reports/farmerreport.R")
 # }
+