+++ R/analysis_climmob.R
@@ -4,10 +4,7 @@
 # crowdsourcing trials using Plackett-Luce model
 # ................................................................
 # ................................................................
-# Kaue de Sousa 
-# Updated 12Mar2020
-# .......................................................
-# .......................................................
+
 # Organise the rankings and check for missing data ####
 trait <- pars$chars$char
 trait_full <- pars$chars$char_full
@@ -145,15 +142,50 @@ if (all(expvar != "xinterceptx")) {
 # Make map ####
 
 # Check if lonlat is provided
-grepl("_lon", names(cmdata))
+lon <- grepl("_lon", names(cmdata))
+lat <- grepl("_lat", names(cmdata))
 
-plot_map(x, c("lon","lat"), map.types = "OpenTopoMap")
+geoTRUE <- any(lon) & any(lat)
+
+if (isTRUE(geoTRUE)) {
+  lon <- which(lon)[1]
+  lat <- which(lat)[1]
+  
+  lonlat <- cmdata[,c(lon,lat)]
+  
+  lonlat <- na.omit(lonlat)
+  
+  nlonlat <- dim(lonlat)[[1]]
+  
+  if (nlonlat > 0){
+    
+    names(lonlat) <- c("lon","lat")
+    
+    trial_map <- plot_map(lonlat, c("lon","lat"),
+                          map.types = c("CartoDB.Positron", "OpenTopoMap"))
+    
+    mapshot(trial_map, 
+            url = paste0(getwd(), "/", pathname, projname, "_trial_map.html"),
+            file = paste0(getwd(), "/", pathname, projname, "_trial_map.png"))
+    
+    trial_map_statement <- c("The map below shows the distribution of the trials in this project.")
+                             
+  
+  }
+  
+}
+
+if (isFALSE(geoTRUE)) {
+  
+  trial_map_statement <- ""
+  
+}
 
 
 # .......................................................
 # .......................................................
-# This is Table 01 #
-# Create a table with frequencies where each item was evaluated ####
+# Make Table 1 ####
+# Create a table with frequencies where each item was evaluated #
 itemdata <- cmdata[, grepl("package_item", names(cmdata))]
 
 itemtable <- data.frame(table(unlist(itemdata)))
@@ -193,7 +225,7 @@ if (isTRUE(gender)) {
 
 itemtable$Abbreviation <- gosset:::.reduce(as.character(itemtable[,Option]))
 
-itemtable <- itemtable[union(c(Option,"Abbreviation"), names(itemtable))]
+itemtable <- itemtable[union(c(Option, "Abbreviation"), names(itemtable))]
 
 # .......................................................
 # .......................................................
@@ -251,8 +283,8 @@ fav2$fav_score <- round(fav2$fav_score, 1)
 
 fav2 <- fav2[,-which(grepl("wins", names(fav2)))]
 
-names(fav2) <- c(Option,"N","Top Ranked","Bottom Ranked",
-                 "Net Favourability Score")
+names(fav2) <- c(Option,"N","Top Ranked",
+                 "Bottom Ranked", "Net Favourability Score")
 
 # Contest Plots
 cont1 <- summarise_dominance(R)
@@ -265,7 +297,9 @@ cont2 <- summarise_victories(R)
 # this assess how the other traits agreed with the overall preference
 # build rankings for the other characteristics
 other_traits <- trait[-1]
+
 other_traits_full <- trait_full[-1]
+
 if (length(other_traits) > 0) {
   
   # filter cmdata so it matches the dims in all traits
@@ -464,14 +498,17 @@ coefs <- as.data.frame(coefs)
 # add item names as rows
 rownames(coefs) <- rownames(qvcalc(mod_overall)[[2]])
 # set col names with title case traits
-names(coefs) <- c("Overall", ClimMobTools:::.title_case(other_traits))
+otf <- gsub(" ","_", other_traits_full)
+names(coefs) <- c("Overall", ClimMobTools:::.title_case(otf))
 
 # .......................................................
 # .......................................................
 # Compute partial least squares ####
+arrows <- data.frame()
+scores <- data.frame()
 
 if (length(other_traits) > 0) {
-  fml <- paste("Overall ~ ", paste(ClimMobTools:::.title_case(other_traits), collapse = " + "))
+  fml <- paste("Overall ~ ", paste(ClimMobTools:::.title_case(otf), collapse = " + "))
   fml <- as.formula(fml)
   
   m2 <- plsr(fml,
@@ -479,11 +516,9 @@ if (length(other_traits) > 0) {
              validation = "LOO", 
              jackknife = TRUE)
   
-  arrows <- NULL
-  scores <- NULL
   if (ncol(m2$projection) > 1 ) {
     arrows <- data.frame((m2$projection)[,1:2],
-                         trait = other_traits,
+                         trait = otf,
                          x0 = 0, 
                          y0 = 0, 
                          stringsAsFactors = FALSE)
@@ -497,7 +532,44 @@ if (length(other_traits) > 0) {
                  intercept = FALSE)$val)
   
   adjCV <- m2$validation$adj
+  
   nc <- which(adjCV == min(adjCV))
+}
+
+if (dim(arrows)[[1]] > 0) {
+  
+  pls_plot <- ggplot(data = arrows,
+                     aes(y = Comp.2, 
+                         x = Comp.1, 
+                         label = trait, 
+                         yend = y0, 
+                         xend = x0)) +
+    geom_hline(yintercept = 0) + 
+    geom_vline(xintercept = 0) +
+    geom_segment(col = "red", arrow = arrow(length = unit(0.5, "cm"), ends = "first" )) +
+    geom_abline(linetype = 2, col = "gray50", slope = (yve[2] - yve[1]) / yve[1], intercept = 0) +
+    geom_text_repel(fontface = 2, size = 4, col = "red") +
+    geom_text(data = scores, 
+              aes(y = Comp.2,
+                  x = Comp.1, 
+                  label = var), 
+              inherit.aes = FALSE, fontface = 2, size = 3) +
+    theme_bw() +
+    theme(panel.grid.minor = element_blank()) +
+    labs(x = paste0("PC1 ", round(yve[1] * 100, 2), "%"),
+         y = paste0("PC2 ", round((yve[2]-yve[1]) * 100, 2),"%"))
+  
+  pls_summary_line <- paste("The first two components recombining the specific characteristics",
+  "are able to explain", paste0(round(yve[2] * 100, 2), "%"), "of the variability in the overall performance.",
+  "The dashed line represents the overall performance, with an increase in performance as the x and y increase.",
+  "The", option, "positioned close to the dashed line will be performing equally across all characteristics;",
+  "the", option, "positioned further away from the dashed line, on either side, will have varying performance",
+  "in different characteristics. Better performance in characteristics will correspond with arrows pointing",
+  "in the direction away from the dashed line and worse performance in characteristics directed on the opposite side.") 
+}
+
+if (dim(arrows)[[1]] == 0) {
+  pls_summary_line<-"However, partial least squares analysis was not possible with this project data."
 }
 
 # .......................................................
@@ -545,12 +617,25 @@ if (isFALSE(overallVSlocal)) {
 
 # data frame of explanatory variables
 Gdata <- as.data.frame(cmdata[keep, expvar_list$expvar], stringsAsFactors = TRUE)
+nvar <- length(expvar_list$expvar)
+# setup the variables
+Gdata[1:nvar] <- lapply(Gdata[1:nvar], function(x){
+  if(is.numeric(x)) {
+    x <- round(x, 3)
+  }
+  
+  if(is.character(x)) {
+    x <- as.factor(x)
+  }
+  return(x)
+})
+
 names(Gdata) <- expvar_list$expvar_full
 Gdata <- cbind(G, Gdata)
 
 tree_f <- pltree(G ~ .,
                  data = Gdata, 
-                 minsize = minsplit, 
+                 minsize = minsplit,
                  alpha = sig_level)
 
 
@@ -613,11 +698,14 @@ for(j in seq_along(tree_f)){
     })[[1]]
   
   if (length(zzz) > 0) {
+    x <- data.frame(Node = j,
+                    t(nodeapply(tree_f, j, function(n){
+                      info_node(n)$test
+                    })[[1]]))
+      
+    x$p.value <- round(x$p.value, 3)
     
-    outtabs[[j]] <- data.frame(Node = j,
-                               t(nodeapply(tree_f, j, function(n){
-                                 info_node(n)$test
-                                 })[[1]]))
+    outtabs[[j]] <- x
     
     outtabs[[j]]$p <- format.pval(outtabs[[j]]$p.value)
   }
@@ -630,17 +718,53 @@ for(j in seq_along(tree_f)){
   }
 }
 
+# Built table with p-values per node
+pval_nodes <- data.frame()
+
+for (i in seq_along(outtabs)) {
+  if (ncol(outtabs[[i]]) > 3) {
+    ot <- data.frame(Covariate = rownames(outtabs[[i]]),
+                     outtabs[[i]])
+    pval_nodes <- rbind(pval_nodes, ot)
+  }
+}
+
+if (dim(pval_nodes)[[2]] > 0) {
+  
+  pval_nodes <- pval_nodes[,c("Covariate", "Node", "p")]
+  
+  row.names(pval_nodes) <- NULL
+  
+  # put bold in significant values 
+  putbold <- which(pval_nodes$p <= sig_level)
+  
+  for(i in seq_along(putbold)) {
+    index_i <- putbold[i]
+    
+    pval_nodes[index_i, ] <- paste0("**", pval_nodes[index_i, ], "**")
+    
+  }
+  
+  nodemessage <- ""
+  
+}
+
+if (dim(pval_nodes)[[2]] == 0) {
+  nodemessage <- "No covariates were able to be assessed statistically."
+}
+
 # ....................................................................
 # ....................................................................
 # Build headline summaries ####
 
 siglist <- NULL
-for(i in 1:length(outtabs)){
-  if (ncol(outtabs[[i]]) > 3) {
+for(i in seq_along(outtabs)){
+  if (dim(outtabs[[i]])[[2]] > 3) {
     siglist <- c(siglist,
-                 rownames(outtabs[[i]])[outtabs[[i]]$p.value<0.05])
+                 rownames(outtabs[[i]])[outtabs[[i]]$p.value < sig_level])
   }
 }
+
 siglist <- unique(siglist)
 
 ps <- fullanova[2, 5]
@@ -660,7 +784,7 @@ if (length(other_traits) > 0) {
     
     ps <- c(ps, anovas[[i]][2,5])
     
-    if(ps[i]<0.05){
+    if(ps[i] < sig_level){
       bests<-c(bests,paste(summaries[[i]]$term[grep("a",summaries[[i]]$.group)],collapse=", "))
       worsts<-c(worsts,paste(rev(summaries[[i]]$term[grep(summaries[[i]]$.group[nrow(summaries[[i]])],
                                                           summaries[[i]]$.group)]),collapse=", "))
@@ -703,7 +827,7 @@ rownames(uni_sum) <- NULL
 
 
 # define height of plots based on items
-favplot_h <- nitems * 0.6
-contest_h <- nitems * 0.6 * 2
-agreem_h <- ntrait * 1.3
-multcomp_h <- nitems * 0.8 
+favplot_h <- nitems * 0.4
+contest_h <- nitems * 0.4 * 2
+agreem_h <- ntrait * 0.9
+multcomp_h <- nitems * 0.6